agents/introvert-ally-api/requirements.txt
functions-framework
flask
google-cloud-aiplatform
python-dotenv
# google-cloud-discoveryengine # Descomenta si usas otras partes del SDK